+++ src/components/KMultiselect/KMultiselectItem.vue
@@ -65,6 +65,7 @@ const handleClick = (): void => {
     background-color: var(--kui-color-background, $kui-color-background);
     border: none;
     display: flex;
+    outline-offset: -1px;
     padding: var(--kui-space-0, $kui-space-0);
     text-align: left;
     width: 100%;