+++ src/types/clipboard-provider.ts
@@ -1,0 +1,8 @@
+import type { copyTextToClipboard } from '@/utilities/copyTextToClipboard'
+
+export interface ClipboardProviderSlots {
+  /**
+   * Slot for providing `copyToClipboard` method to the children.
+   */
+  default?(props: { copyToClipboard: (arg: string | undefined) => ReturnType<typeof copyTextToClipboard> }): any
+}